requirements.txt
ultralytics
flask
flask-cors
requests
python-dotenv
torch
torchvision
gunicorn
python-multipart
opencv-python